Add Path

The Path Layer takes in lists of coordinate points and renders them as extruded lines with mitering.

add_path(map, data = get_map_data(map), polyline = NULL,
  stroke_colour = NULL, stroke_width = NULL, stroke_opacity = NULL,
  tooltip = NULL, layer_id = NULL, id = NULL,
  auto_highlight = FALSE, highlight_colour = "#AAFFFFFF",
  palette = "viridis", na_colour = "#808080FF", legend = FALSE,
  legend_options = NULL, update_view = TRUE, focus_layer = FALSE,
  transitions = NULL)

Arguments

map
a mapdeck map object

data
data to be used in the layer. All coordinates are expected to be in Web Mercator Projection

polyline
column of data containing the polylines

stroke_colour
variable of data or hex colour for the stroke. transition enabled

stroke_width
width of the stroke. transition enabled

stroke_opacity
value between 1 and 255. Either a string specifying the column of data containing the stroke opacity of each shape, or a value between 1 and 255 to be applied to all the shapes

tooltip
variable of data containing text or HTML to render as a tooltip

layer_id
single value specifying an id for the layer. Use this value to distinguish between shape layers of the same type. Layers with the same id are likely to conflict and not plot correctly

id
an id value in data to identify layers when interacting in Shiny apps.

auto_highlight
logical indicating if the shape under the mouse should auto-highlight

highlight_colour
hex string colour to use for highlighting. Must contain the alpha component.

palette
string or matrix. String is either one of "viridis","inferno", "magma","plasma" or "cividis". A matrix is a 3 or 4 column numeric matrix of values between [0, 255], where the 4th column represents the alpha.

na_colour
hex string colour to use for NA values

legend
either a logical indiciating if the legend(s) should be displayed, or a named list indicating which colour attributes should be included in the legend.

legend_options
A list of options for controlling the legend.

update_view
logical indicating if the map should update the bounds to include this layer

focus_layer
logical indicating if the map should update the bounds to only include this layer

transitions
list specifying the duration of transitions.

Details

add_path supports LINESTRING and MULTILINESTRING sf objects

transitions

The transitions argument lets you specify the time it will take for the shapes to transition from one state to the next. Only works in an interactive environment (Shiny) and on WebGL-2 supported browsers and hardware.

The time is in milliseconds

Available transitions for path

list( path = 0, stroke_colour = 0, stroke_width = 0 )

data

If the data is a simple feature object, the geometry column is automatically detected. If the sf object contains more than one geometry column and you want to use a specific one, you'll need to set the active geometry using sf::st_geometry( x ) <- "your_column" , where "your_column" is the name of the column you're activating. See ?sf::st_geometry

legend

The legend_options can be used to control the appearance of the legend. This should be a named list, where the names are one of

css - a string of valid css for controlling the appearance of the legend

title - a string to use for the title of the legend

If the layer allows different fill and stroke colours, you can use different options for each. See examples in add_arc.

id

The id is returned to your R session from an interactive shiny environment by observing layer clicks. This is useful for returning the data.frame row relating to the cliked shape.

From within a shiny server you would typically use observeEvent({input$map_arc_click}), where 'map' is the map_id supplied to mapdeckOutput(), and 'arc' is the layer you are clicking on

Examples

## You need a valid access token from Mapbox
key <- 'abc'

mapdeck(
  token = key
  , style = 'mapbox://styles/mapbox/dark-v9'
  , location = c(145, -37.8)
  , zoom = 10) %>%
  add_path(
    data = roads
    , stroke_colour = "RIGHT_LOC"
    , layer_id = "path_layer"
    , tooltip = "ROAD_NAME"
    , auto_highlight = TRUE
    , legend = T
  )