Add Polygon

The Polygon Layer renders filled and/or stroked polygons. If using sf objects only POLYGONs are supported, MULTIPOLYGONs are ignored.

add_polygon(map, data = get_map_data(map), polyline = NULL,
  stroke_colour = NULL, stroke_width = NULL, fill_colour = NULL,
  fill_opacity = NULL, elevation = NULL, tooltip = NULL,
  auto_highlight = FALSE, highlight_colour = "#AAFFFFFF",
  light_settings = list(), layer_id = NULL, id = NULL,
  palette = "viridis", na_colour = "#808080FF", legend = FALSE,
  legend_options = NULL, update_view = TRUE, focus_layer = FALSE,
  transitions = NULL)

Arguments

map
a mapdeck map object

data
data to be used in the layer

polyline
column of data containing the polylines

stroke_colour
variable of data or hex colour for the stroke. transition enabled

stroke_width
width of the stroke. transition enabled

fill_colour
column of data or hex colour for the fill colour. transition enabled

fill_opacity
value between 1 and 255. Either a string specifying the column of data containing the fill opacity of each shape, or a value between 1 and 255 to be applied to all the shapes

elevation
the height the polygon extrudes from the map. transition enabled

tooltip
variable of data containing text or HTML to render as a tooltip

auto_highlight
logical indicating if the shape under the mouse should auto-highlight

highlight_colour
hex string colour to use for highlighting. Must contain the alpha component.

light_settings
list of light setting parameters. See light_settings

layer_id
single value specifying an id for the layer. Use this value to distinguish between shape layers of the same type. Layers with the same id are likely to conflict and not plot correctly

id
an id value in data to identify layers when interacting in Shiny apps.

palette
string or matrix. String is either one of "viridis","inferno", "magma","plasma" or "cividis". A matrix is a 3 or 4 column numeric matrix of values between [0, 255], where the 4th column represents the alpha.

na_colour
hex string colour to use for NA values

legend
either a logical indiciating if the legend(s) should be displayed, or a named list indicating which colour attributes should be included in the legend.

legend_options
A list of options for controlling the legend.

update_view
logical indicating if the map should update the bounds to include this layer

focus_layer
logical indicating if the map should update the bounds to only include this layer

transitions
list specifying the duration of transitions.

Details

add_polygon supports POLYGON and MULTIPOLYGON sf objects

transitions

The transitions argument lets you specify the time it will take for the shapes to transition from one state to the next. Only works in an interactive environment (Shiny) and on WebGL-2 supported browsers and hardware.

The time is in milliseconds

Available transitions for polygon

list( polygon = 0, fill_colour = 0, stroke_colour = 0, stroke_width = 0, elevation = 0 )

legend

The legend_options can be used to control the appearance of the legend. This should be a named list, where the names are one of

css - a string of valid css for controlling the appearance of the legend

title - a string to use for the title of the legend

If the layer allows different fill and stroke colours, you can use different options for each. See examples in add_arc.

id

The id is returned to your R session from an interactive shiny environment by observing layer clicks. This is useful for returning the data.frame row relating to the cliked shape.

From within a shiny server you would typically use observeEvent({input$map_arc_click}), where 'map' is the map_id supplied to mapdeckOutput(), and 'arc' is the layer you are clicking on

Examples

## You need a valid access token from Mapbox
key <- 'abc'

df <- melbourne
df$elevation <- sample(100:5000, size = nrow(df))
df$info <- paste0("<b>SA2 - </b><br>",df$SA2_NAME)

mapdeck(
  token = key
  , style = mapdeck_style('dark')
  , location = c(145, -38)
  , zoom = 8
  ) %>%
  add_polygon(
    data = df
    , polyline = "geometry"
    , layer = "polygon_layer"
    , fill_colour = "SA2_NAME",
    , elevation = "elevation"
    , stroke_width = 0
    , tooltip = 'info'
    , legend = T
  )

library(sf)
library(geojsonsf)

sf <- geojsonsf::geojson_sf("https://symbolixau.github.io/data/geojson/SA2_2016_VIC.json")

mapdeck(
  token = key
  , style = mapdeck_style('dark')
) %>%
  add_polygon(
    data = sf
    , layer = "polygon_layer"
    , fill_colour = "SA2_NAME16"
  )